+++ program/src/state.rs
@@ -40,6 +40,15 @@ use crate::{
     VALIDATOR_STAKE_ACCOUNT, VALIDATOR_UNSTAKE_ACCOUNT,
 };
 
+mod validator;
+pub use validator::Validator;
+
+mod validator_perf;
+pub use validator_perf::{Criteria, OffchainValidatorPerf, ValidatorPerf};
+
+mod maintainer;
+pub use maintainer::Maintainer;
+
 /// Types of list entries
 /// Uninitialized should always be a first enum field as it catches empty list data errors
 #[derive(Clone, Debug, PartialEq, Eq, BorshDeserialize, BorshSerialize, Serialize, BorshSchema)]
@@ -320,346 +329,6 @@ impl ValidatorList {
     }
 }
 
-/// NOTE: ORDER IS VERY IMPORTANT HERE, PLEASE DO NOT RE-ORDER THE FIELDS UNLESS
-/// THERE'S AN EXTREMELY GOOD REASON.
-///
-/// To save on BPF instructions, the serialized bytes are reinterpreted with an
-/// unsafe pointer cast, which means that this structure cannot have any
-/// undeclared alignment-padding in its representation.
-#[repr(C)]
-#[derive(Clone, Debug, Eq, PartialEq, BorshDeserialize, BorshSerialize, BorshSchema, Serialize)]
-pub struct Validator {
-    /// Validator vote account address.
-    /// Do not reorder this field, it should be first in the struct
-    #[serde(serialize_with = "serialize_b58")]
-    #[serde(rename = "pubkey")]
-    pub vote_account_address: Pubkey,
-
-    /// Seeds for active stake accounts.
-    pub stake_seeds: SeedRange,
-    /// Seeds for inactive stake accounts.
-    pub unstake_seeds: SeedRange,
-
-    /// Sum of the balances of the stake accounts and unstake accounts.
-    pub stake_accounts_balance: Lamports,
-
-    /// Sum of the balances of the unstake accounts.
-    pub unstake_accounts_balance: Lamports,
-
-    /// Effective stake balance is stake_accounts_balance - unstake_accounts_balance.
-    /// The result is stored on-chain to optimize compute budget
-    pub effective_stake_balance: Lamports,
-
-    /// Controls if a validator is allowed to have new stake deposits.
-    /// When removing a validator, this flag should be set to `false`.
-    pub active: bool,
-}
-
-/// NOTE: ORDER IS VERY IMPORTANT HERE, PLEASE DO NOT RE-ORDER THE FIELDS UNLESS
-/// THERE'S AN EXTREMELY GOOD REASON.
-///
-/// To save on BPF instructions, the serialized bytes are reinterpreted with an
-/// unsafe pointer cast, which means that this structure cannot have any
-/// undeclared alignment-padding in its representation.
-#[repr(C)]
-#[derive(Clone, Debug, Eq, PartialEq, BorshDeserialize, BorshSerialize, BorshSchema, Serialize)]
-pub struct OffchainValidatorPerf {
-    /// The epoch in which the off-chain part
-    /// of the validator's performance was computed.
-    pub updated_at: Epoch,
-
-    /// The number of slots the validator has produced in the last epoch.
-    pub block_production_rate: u64,
-
-    /// Ratio of successful votes to total votes.
-    pub vote_success_rate: u64,
-
-    /// Ratio of how long the validator has been available to the total time in the epoch.
-    pub uptime: u64,
-}
-
-/// NOTE: ORDER IS VERY IMPORTANT HERE, PLEASE DO NOT RE-ORDER THE FIELDS UNLESS
-/// THERE'S AN EXTREMELY GOOD REASON.
-///
-/// To save on BPF instructions, the serialized bytes are reinterpreted with an
-/// unsafe pointer cast, which means that this structure cannot have any
-/// undeclared alignment-padding in its representation.
-#[repr(C)]
-#[derive(
-    Clone, Debug, Default, Eq, PartialEq, BorshDeserialize, BorshSerialize, BorshSchema, Serialize,
-)]
-pub struct ValidatorPerf {
-    /// The associated validator's vote account address.
-    /// It might not be present in the validator list.
-    /// Do not reorder this field, it should be first in the struct
-    #[serde(serialize_with = "serialize_b58")]
-    #[serde(rename = "pubkey")]
-    pub validator_vote_account_address: Pubkey,
-
-    /// The commission is updated at its own pace.
-    pub commission: u8,
-    pub commission_updated_at: Epoch,
-
-    /// The off-chain part of the validator's performance, if available.
-    pub rest: Option<OffchainValidatorPerf>,
-}
-
-impl ValidatorPerf {
-    /// True only if these metrics meet the criteria.
-    pub fn meets_criteria(&self, criteria: &Criteria) -> bool {
-        self.commission <= criteria.max_commission
-            && self.rest.as_ref().map_or(true, |perf| {
-                perf.vote_success_rate >= criteria.min_vote_success_rate
-                    && perf.block_production_rate >= criteria.min_block_production_rate
-                    && perf.uptime >= criteria.min_uptime
-            })
-    }
-}
-
-/// NOTE: ORDER IS VERY IMPORTANT HERE, PLEASE DO NOT RE-ORDER THE FIELDS UNLESS
-/// THERE'S AN EXTREMELY GOOD REASON.
-///
-/// To save on BPF instructions, the serialized bytes are reinterpreted with an
-/// unsafe pointer cast, which means that this structure cannot have any
-/// undeclared alignment-padding in its representation.
-#[repr(C)]
-#[derive(
-    Clone, Default, Debug, Eq, PartialEq, BorshDeserialize, BorshSerialize, BorshSchema, Serialize,
-)]
-pub struct Maintainer {
-    /// Address of maintainer account.
-    /// Do not reorder this field, it should be first in the struct
-    #[serde(serialize_with = "serialize_b58")]
-    pub pubkey: Pubkey,
-}
-
-impl Validator {
-    /// Return the balance in only the stake accounts, excluding the unstake accounts.
-    pub fn compute_effective_stake_balance(&self) -> Lamports {
-        (self.stake_accounts_balance - self.unstake_accounts_balance)
-            .expect("Unstake balance cannot exceed the validator's total stake balance.")
-    }
-
-    pub fn observe_balance(observed: Lamports, tracked: Lamports, info: &str) -> ProgramResult {
-        if observed < tracked {
-            msg!(
-                "{}: observed balance of {} is less than tracked balance of {}.",
-                info,
-                observed,
-                tracked
-            );
-            msg!("This should not happen, aborting ...");
-            return Err(LidoError::ValidatorBalanceDecreased.into());
-        }
-        Ok(())
-    }
-
-    pub fn has_stake_accounts(&self) -> bool {
-        self.stake_seeds.begin != self.stake_seeds.end
-    }
-    pub fn has_unstake_accounts(&self) -> bool {
-        self.unstake_seeds.begin != self.unstake_seeds.end
-    }
-
-    pub fn check_can_be_removed(&self) -> Result<(), LidoError> {
-        if self.active {
-            return Err(LidoError::ValidatorIsStillActive);
-        }
-        if self.has_stake_accounts() {
-            return Err(LidoError::ValidatorShouldHaveNoStakeAccounts);
-        }
-        if self.has_unstake_accounts() {
-            return Err(LidoError::ValidatorShouldHaveNoUnstakeAccounts);
-        }
-        // If not, this is a bug.
-        assert_eq!(self.stake_accounts_balance, Lamports(0));
-        Ok(())
-    }
-
-    pub fn show_removed_error_msg(error: &Result<(), LidoError>) {
-        if let Err(err) = error {
-            match err {
-                LidoError::ValidatorIsStillActive => {
-                    msg!(
-                                "Refusing to remove validator because it is still active, deactivate it first."
-                            );
-                }
-                LidoError::ValidatorHasUnclaimedCredit => {
-                    msg!(
-                        "Validator still has tokens to claim. Reclaim tokens before removing the validator"
-                    );
-                }
-                LidoError::ValidatorShouldHaveNoStakeAccounts => {
-                    msg!("Refusing to remove validator because it still has stake accounts, unstake them first.");
-                }
-                LidoError::ValidatorShouldHaveNoUnstakeAccounts => {
-                    msg!("Refusing to remove validator because it still has unstake accounts, withdraw them first.");
-                }
-                _ => {
-                    msg!("Invalid error when removing a validator: shouldn't happen.");
-                }
-            }
-        }
-    }
-
-    pub fn find_stake_account_address_with_authority(
-        &self,
-        program_id: &Pubkey,
-        solido_account: &Pubkey,
-        authority: &[u8],
-        seed: u64,
-    ) -> (Pubkey, u8) {
-        let seeds = [
-            &solido_account.to_bytes(),
-            &self.vote_account_address.to_bytes(),
-            authority,
-            &seed.to_le_bytes()[..],
-        ];
-        Pubkey::find_program_address(&seeds, program_id)
-    }
-
-    pub fn find_stake_account_address(
-        &self,
-        program_id: &Pubkey,
-        solido_account: &Pubkey,
-        seed: u64,
-        stake_type: StakeType,
-    ) -> (Pubkey, u8) {
-        let authority = match stake_type {
-            StakeType::Stake => VALIDATOR_STAKE_ACCOUNT,
-            StakeType::Unstake => VALIDATOR_UNSTAKE_ACCOUNT,
-        };
-        self.find_stake_account_address_with_authority(program_id, solido_account, authority, seed)
-    }
-
-    /// Get stake account address that should be merged into another right after creation.
-    /// This function should be used to create temporary stake accounts
-    /// tied to the epoch that should be merged into another account and destroyed
-    /// after a transaction. So that each epoch would have a different
-    /// generation of stake accounts. This is done for security purpose
-    pub fn find_temporary_stake_account_address(
-        &self,
-        program_id: &Pubkey,
-        solido_account: &Pubkey,
-        seed: u64,
-        epoch: Epoch,
-    ) -> (Pubkey, u8) {
-        let authority = [VALIDATOR_STAKE_ACCOUNT, &epoch.to_le_bytes()[..]].concat();
-        self.find_stake_account_address_with_authority(program_id, solido_account, &authority, seed)
-    }
-
-    /// Mark the validator as active so that they could receive new stake.
-    pub fn activate(&mut self) {
-        self.active = true;
-    }
-
-    /// Mark the validator as inactive so that no new stake can be delegated to it,
-    /// and the existing stake shall be unstaked by the maintainer.
-    pub fn deactivate(&mut self) {
-        self.active = false;
-    }
-}
-
-impl Sealed for Validator {}
-
-impl Pack for Validator {
-    const LEN: usize = 89;
-    fn pack_into_slice(&self, data: &mut [u8]) {
-        let mut data = data;
-        BorshSerialize::serialize(&self, &mut data).unwrap();
-    }
-    fn unpack_from_slice(src: &[u8]) -> Result<Self, ProgramError> {
-        let unpacked = Self::try_from_slice(src)?;
-        Ok(unpacked)
-    }
-}
-
-impl Default for Validator {
-    fn default() -> Self {
-        Validator {
-            stake_seeds: SeedRange { begin: 0, end: 0 },
-            unstake_seeds: SeedRange { begin: 0, end: 0 },
-            stake_accounts_balance: Lamports(0),
-            unstake_accounts_balance: Lamports(0),
-            effective_stake_balance: Lamports(0),
-            active: true,
-            vote_account_address: Pubkey::default(),
-        }
-    }
-}
-
-impl ListEntry for Validator {
-    const TYPE: AccountType = AccountType::Validator;
-
-    fn new(vote_account_address: Pubkey) -> Self {
-        Self {
-            vote_account_address,
-            ..Default::default()
-        }
-    }
-
-    fn pubkey(&self) -> &Pubkey {
-        &self.vote_account_address
-    }
-}
-
-impl ValidatorPerf {}
-
-impl Sealed for ValidatorPerf {}
-
-impl Pack for ValidatorPerf {
-    const LEN: usize = 64;
-    fn pack_into_slice(&self, data: &mut [u8]) {
-        let mut data = data;
-        BorshSerialize::serialize(&self, &mut data).unwrap();
-    }
-    fn unpack_from_slice(src: &[u8]) -> Result<Self, ProgramError> {
-        let unpacked = Self::try_from_slice(src)?;
-        Ok(unpacked)
-    }
-}
-
-impl ListEntry for ValidatorPerf {
-    const TYPE: AccountType = AccountType::ValidatorPerf;
-
-    fn new(validator_vote_account_address: Pubkey) -> Self {
-        Self {
-            validator_vote_account_address,
-            ..Default::default()
-        }
-    }
-
-    fn pubkey(&self) -> &Pubkey {
-        &self.validator_vote_account_address
-    }
-}
-
-impl Sealed for Maintainer {}
-
-impl Pack for Maintainer {
-    const LEN: usize = PUBKEY_BYTES;
-    fn pack_into_slice(&self, data: &mut [u8]) {
-        let mut data = data;
-        BorshSerialize::serialize(&self, &mut data).unwrap();
-    }
-    fn unpack_from_slice(src: &[u8]) -> Result<Self, ProgramError> {
-        let unpacked = Self::try_from_slice(src)?;
-        Ok(unpacked)
-    }
-}
-
-impl ListEntry for Maintainer {
-    const TYPE: AccountType = AccountType::Maintainer;
-
-    fn new(pubkey: Pubkey) -> Self {
-        Self { pubkey }
-    }
-
-    fn pubkey(&self) -> &Pubkey {
-        &self.pubkey
-    }
-}
-
 /// The exchange rate used for deposits and rewards distribution.
 ///
 /// The exchange rate of SOL to stSOL is determined by the SOL balance of
@@ -780,53 +449,6 @@ impl ExchangeRate {
         // dimensionless, but in this case `rate` has dimensions SOL/stSOL, so
         // we need to re-wrap the result in the right type.
         Ok((amount * rate).map(|x| Lamports(x.0))?)
-    }
-}
-
-/// Each field is an optimum for a metric.
-/// If a validator has a value for a metric that does not meet the threshold,
-/// then the validator gets deactivated.
-///
-#[repr(C)]
-#[derive(Clone, Debug, BorshDeserialize, BorshSerialize, BorshSchema, Eq, PartialEq, Serialize)]
-pub struct Criteria {
-    /// If a validator has the commission higher than this, then it gets deactivated.
-    pub max_commission: u8,
-
-    /// If a validator has `block_production_rate` lower than this, then it gets deactivated.
-    pub min_block_production_rate: u64,
-
-    /// If a validator has `vote_success_rate` lower than this, then it gets deactivated.
-    pub min_vote_success_rate: u64,
-
-    /// If a validator has the uptime lower than this, then it gets deactivated.
-    pub min_uptime: u64,
-}
-
-impl Default for Criteria {
-    fn default() -> Self {
-        Self {
-            max_commission: 100,
-            min_vote_success_rate: 0,
-            min_block_production_rate: 0,
-            min_uptime: 0,
-        }
-    }
-}
-
-impl Criteria {
-    pub fn new(
-        max_commission: u8,
-        min_vote_success_rate: u64,
-        min_block_production_rate: u64,
-        min_uptime: u64,
-    ) -> Self {
-        Self {
-            max_commission,
-            min_vote_success_rate,
-            min_block_production_rate,
-            min_uptime,
-        }
     }
 }
